+++ xllm/core/framework/request/request_output.h
@@ -28,13 +28,13 @@ limitations under the License.
 namespace xllm {
 struct Usage {
   // the number of tokens in the prompt.
-  size_t num_prompt_tokens = 0;
+  int32_t num_prompt_tokens = 0;
 
   // the number of tokens in the generated completion.
-  size_t num_generated_tokens = 0;
+  int32_t num_generated_tokens = 0;
 
   // the total number of tokens used in the request (prompt + completion).
-  size_t num_total_tokens = 0;
+  int32_t num_total_tokens = 0;
 };
 
 struct LogProbData {
